Moving generic tracing to tracing instead of jaeger
