+++ tracing/pg/instrument.go
@@ -20,19 +20,19 @@
  * CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-package gorp
+package pg
 
 import (
 	"context"
-	"regexp"
 	"strings"
 
+	"github.com/go-pg/pg/orm"
 	"github.com/opentracing/opentracing-go"
-	"github.com/topfreegames/extensions/jaeger"
+	"github.com/topfreegames/extensions/tracing"
 )
 
-// Trace wraps a gorp database call and reports it to Jaeger
-func Trace(ctx context.Context, name string, query string, next func() error) {
+// Trace wraps a go-pg query and reports it to tracing
+func Trace(ctx context.Context, query interface{}, next func() error) {
 	var parent opentracing.SpanContext
 
 	if ctx == nil {
@@ -43,30 +43,33 @@ func Trace(ctx context.Context, name string, query string, next func() error) {
 		parent = span.Context()
 	}
 
-	operationName := "SQL " + parse(query)
+	statement := "UNKNOWN STATEMENT"
+	if val, ok := query.(string); ok {
+		statement = val
+	} else if val, ok := query.(orm.QueryAppender); ok {
+		if statementBytes, err := val.AppendQuery(nil); err == nil {
+			statement = string(statementBytes)
+		}
+	}
+
+	operation := strings.Fields(statement)[0]
+	operationName := "SQL " + operation
 	reference := opentracing.ChildOf(parent)
 	tags := opentracing.Tags{
-		"db.instance":  name,
-		"db.statement": query,
+		"db.operation": operation,
 		"db.type":      "postgres",
+		"db.statement": statement,
 
 		"span.kind": "client",
 	}
 
 	span := opentracing.StartSpan(operationName, reference, tags)
 	defer span.Finish()
-	defer jaeger.LogPanic(span)
+	defer tracing.LogPanic(span)
 
 	err := next()
 	if err != nil {
 		message := err.Error()
-		jaeger.LogError(span, message)
+		tracing.LogError(span, message)
 	}
-}
-
-func parse(query string) string {
-	re := regexp.MustCompile("\\s+")
-	array := re.Split(" "+query, 3)
-	command := array[1]
-	return strings.ToUpper(command)
 }
