+++ util/version.go
@@ -23,4 +23,4 @@
 package util
 
 //Version identifies extensions package version
-const Version = "6.4.5"
+const Version = "7.0.0"
